# fastapi_backend/app/routes/items.py
from uuid import UUID

from fastapi import APIRouter, Depends, HTTPException, Query
from fastapi_pagination import Page, Params
from fastapi_pagination.ext.sqlalchemy import apaginate
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.future import select

from app.database import User, get_async_session
from app.models import Item
from app.schemas import ItemRead, ItemCreate
from app.users import current_active_user

router = APIRouter(tags=["item"])


def transform_items(items):
    return [ItemRead.model_validate(item) for item in items]


@router.get(
    "/",
    summary="Paginated list (authenticated)",
    response_model=Page[ItemRead],
)
async def read_item(
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
    page: int = Query(1, ge=1, description="Page number"),
    size: int = Query(10, ge=1, le=100, description="Page size"),
):
    """Return a paginated list of the current user's items.

    Requires POST /auth/jwt/login.
    """
    params = Params(page=page, size=size)
    query = select(Item).filter(Item.user_id == user.id)
    return await apaginate(db, query, params, transformer=transform_items)


@router.post(
    "/",
    summary="Create item for current user",
    response_model=ItemRead,
)
async def create_item(
    item: ItemCreate,
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):
    """Create an item for the current user.

    Requires POST /auth/jwt/login.
    """
    db_item = Item(**item.model_dump(), user_id=user.id)
    db.add(db_item)
    await db.commit()
    await db.refresh(db_item)
    return db_item


@router.delete("/{item_id}", summary="Delete own item")
async def delete_item(
    item_id: UUID,
    db: AsyncSession = Depends(get_async_session),
    user: User = Depends(current_active_user),
):
    """Delete the current user's item by id.

    Requires POST /auth/jwt/login. Returns 404 if missing or not owned.
    """
    result = await db.execute(
        select(Item).filter(Item.id == item_id, Item.user_id == user.id)
    )
    item = result.scalars().first()

    if not item:
        raise HTTPException(status_code=404, detail="Item not found or not authorized")

    await db.delete(item)
    await db.commit()

    return {"message": "Item successfully deleted"}